+++ benches/encrypted_message.rs
@@ -7,8 +7,20 @@ use encrypted_message::{
 use rand::distributions::{Alphanumeric, DistString};
 
 #[derive(Debug, Default)]
-pub struct BenchesConfig;
-impl Config for BenchesConfig {
+pub struct ConfigDeterministic;
+impl Config for ConfigDeterministic {
+    type Strategy = Deterministic;
+
+    fn keys(&self) -> Vec<Secret<[u8; 32]>> {
+        vec![(*b"uuOxfpWgRgIEo3dIrdo0hnHJHF1hntvW").into()]
+    }
+}
+
+#[derive(Debug, Default)]
+pub struct ConfigRandomized;
+impl Config for ConfigRandomized {
+    type Strategy = Randomized;
+
     fn keys(&self) -> Vec<Secret<[u8; 32]>> {
         vec![(*b"uuOxfpWgRgIEo3dIrdo0hnHJHF1hntvW").into()]
     }
@@ -19,15 +31,15 @@ fn encrypted_message(c: &mut Criterion) {
     let payload = black_box(Alphanumeric.sample_string(&mut rand::thread_rng(), 32));
 
     c.bench_function("Encrypt 32-byte payload (Deterministic)", |b| b.iter(|| {
-        EncryptedMessage::<_, Deterministic, BenchesConfig>::encrypt(payload.clone()).unwrap()
+        EncryptedMessage::<_, ConfigDeterministic>::encrypt(payload.clone()).unwrap()
     }));
 
     c.bench_function("Encrypt 32-byte payload (Randomized)", |b| b.iter(|| {
-        EncryptedMessage::<_, Randomized, BenchesConfig>::encrypt(payload.clone()).unwrap()
+        EncryptedMessage::<_, ConfigRandomized>::encrypt(payload.clone()).unwrap()
     }));
 
     c.bench_function("Decrypt 32-byte payload", |b| {
-        let encrypted = EncryptedMessage::<_, Deterministic, BenchesConfig>::encrypt(payload.clone()).unwrap();
+        let encrypted = EncryptedMessage::<_, ConfigRandomized>::encrypt(payload.clone()).unwrap();
         b.iter(|| encrypted.decrypt().unwrap())
     });
 }