+++ src/strategy.rs
@@ -54,14 +54,18 @@ mod tests {
 
     use secrecy::ExposeSecret as _;
 
-    use crate::{config::Config as _, utilities::base64, testing::TestConfig};
+    use crate::{
+        config::Config as _,
+        testing::{TestConfigDeterministic, TestConfigRandomized},
+        utilities::base64,
+    };
 
     mod deterministic {
         use super::*;
 
         #[test]
         fn nonce_is_deterministic() {
-            let key = TestConfig.primary_key();
+            let key = TestConfigDeterministic.primary_key();
             let nonce = Deterministic::generate_nonce_for("rigo is cool".as_bytes(), key.expose_secret());
 
             // Test that the nonce is 12 bytes long.
@@ -78,7 +82,7 @@ mod tests {
         #[test]
         fn nonce_is_randomized() {
             let payload = "much secret much secure".as_bytes();
-            let key = TestConfig.primary_key();
+            let key = TestConfigRandomized.primary_key();
             let first_nonce = Randomized::generate_nonce_for(payload, key.expose_secret());
             let second_nonce = Randomized::generate_nonce_for(payload, key.expose_secret());